+++ todo.md
@@ -1,0 +1,3 @@
+Features to-do:
+- [ ] add config "has been edited" message
+- [ ] add edit logic
